Refactor menu handling by moving makeMenu function to a new file and removing logout button from main screen
