+++ gui/app.go
@@ -44,25 +44,3 @@ func Start(app fyne.App, cfg config.Config) {
 
 	gui.Window.ShowAndRun()
 }
-
-func (g *GUI) makeMenu() *fyne.MainMenu {
-	logout := fyne.NewMenuItem("Log Out", func() {
-		g.User = &database.User{}
-		g.Calendar = &database.Calendar{}
-		g.showLogin()
-	})
-	err := g.checkUserLoggedIn()
-	if err != nil {
-		logout.Disabled = true
-	} else {
-		logout.Disabled = false
-	}
-
-	menu := fyne.NewMenu(
-		"Menu",
-		logout)
-	main := fyne.NewMainMenu(
-		menu,
-	)
-	return main
-}
